# api-docs/swagger.json
{
  "info": {
    "title": "Staffing API",
    "description": "Staffing API Sandbox",
    "version": "1.0.0"
  },
  "schemes": [
    "https"
  ],
  "basePath": "/api",
  "swagger": "2.0",
  "paths": {
...
    "/candidates": {
      "get": {
        "tags": [
          "Candidate"
        ],
        "summary": "List all candidates",
        "description": "List all candidates.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Candidate"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Candidate"
        ],
        "summary": "Create candidate",
        "description": "Create candidate profile with form fields",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "avatar",
            "in": "formData",
            "description": "Candidate avatar",
            "required": false,
            "type": "file"
          },
          {
            "name": "interviewVideo",
            "in": "formData",
            "description": "Candidate interview video",
            "required": false,
            "type": "file"
          },
          {
            "name": "fullname",
            "in": "formData",
            "description": "Full name",
            "required": true,
            "type": "string"
          },
          {
            "name": "location",
            "in": "formData",
            "description": "Location",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in": "formData",
            "description": "Candidate title",
            "required": true,
            "type": "string"
          },
          {
            "name": "roles",
            "in": "formData",
            "description": "List of role ids (Stringified array of roles)",
            "required": true,
            "type": "array"
          },
          {
            "name": "skillset",
            "in": "formData",
            "description": "Stringified array of skillset",
            "required": true,
            "type": "array"
          },
          {
            "name": "pastCompanies",
            "in": "formData",
            "description": "Past companies in which he worked (Stringified array of pastCompanies)",
            "required": true,
            "type": "array"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Candidate"
            }
          },
          "400": {
            "description": "error creating"
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      },
      "put": {
        "tags": [
          "Candidate"
        ],
        "summary": "Update candidate",
        "description": "Update proposal",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Candidate"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
    "/candidates/:id": {
      "get": {
        "tags": [
          "Candidate"
        ],
        "summary": "Get candidate by id",
        "description": "Get candidate by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Candidate"
            }
          },
          "500": {
            "description": "internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "Candidate"
        ],
        "summary": "Delete candidate by id",
        "description": "Delete candidate by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Candidate"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
    "/companies": {
      "get": {
        "tags": [
          "Company"
        ],
        "summary": "List companies by query",
        "description": "List companies by query. if no query is sent, list all companies",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      },
      "put": {
        "tags": [
          "Company"
        ],
        "summary": "Update company info",
        "description": "Update company info and store in edit layer without changing original info",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "_id",
            "type": "string",
            "description": "_id of company document",
            "required": true
          },
          {
            "name": "source",
            "type": "string",
            "description": "Who is editing. It will be user when editing from client forms. Or it can by any scrape source like crunchbase, craft etc",
            "required": true
          },
          {
            "name": "companyName",
            "type": "string"
          },
          {
            "name": "description",
            "type": "string"
          },
          {
            "name": "foundersNames",
            "type": "array",
            "items": {
              "type": "string"
            },
            "in": "formData"
          },
          {
            "name": "categories",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "keyPeople",
            "type": "array",
            "items": {
              "type": "object",
              "schema": {
                "$ref": "#/definitions/KeyPeople"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Company": {
      "type": "object",
      "properties": {
        "companyName": {
          "type": "string"
        },
        "foundersNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "cityHeadQuartersIn": {
          "type": "string"
        },
        "numberOfEmployees": {
          "type": "number"
        },
        "websiteLink": {
          "type": "string"
        },
        "fundingRounds": {
          "type": "number"
        },
        "fundingAmount": {
          "type": "string"
        },
        "lastFundingDate": {
          "type": "string"
        },
        "dateFounded": {
          "type": "string"
        },
        "categories": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "angelId": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "acquisitionsNumber": {
          "type": "string"
        },
        "linkedInURL": {
          "type": "string"
        },
        "twitterURL": {
          "type": "string"
        },
        "facebookURL": {
          "type": "string"
        },
        "youtubeURL": {
          "type": "string"
        },
        "instagramURL": {
          "type": "string"
        },
        "angelURL": {
          "type": "string"
        },
        "crunchBaseURL": {
          "type": "string"
        },
        "craftURL": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "keyPeople": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/KeyPeople"
            }
          }
        },
        "source": {
          "type": "string"
        },
        "logos": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Logos"
            }
          }
        },
        "otherLocations": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "edits": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Edits"
            }
          }
        }
      }
    },
    "SkillSet": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "Candidate": {
      "type": "object",
      "required": [
        "roles"
      ],
      "properties": {
        "_id": {
          "type": "string"
        },
        "avatar": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/Image"
          }
        },
        "interviewVideo": {
          "type": "string"
        },
        "fullname": {
          "type": "imageSchema"
        },
        "location": {
          "type": "string"
        },
        "title": {
          "type": "imageSchema"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "skillset": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/SkillSet"
          }
        },
        "pastCompanies": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Company"
            }
          }
        }
      }
    },
    "KeyPeople": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "linkedIn": {
          "type": "string"
        },
        "facebook": {
          "type": "string"
        },
        "twitter": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "other": {
          "type": "string"
        }
      }
    },
    "Logos": {
      "type": "object",
      "properties": {
        "craftLogo": {
          "type": "string"
        },
        "crunchBaseLogo": {
          "type": "string"
        }
      }
    },
    "Edits": {
      "type": "object",
      "properties": {
        "companyName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "numberOfEmployees": {
          "type": "number"
        },
        "dateFounded": {
          "type": "string"
        },
        "websiteLink": {
          "type": "string"
        },
        "linkedInURL": {
          "type": "string"
        },
        "twitterURL": {
          "type": "string"
        },
        "facebookURL": {
          "type": "string"
        },
        "fundingAmount": {
          "type": "string"
        },
        "fundingRounds": {
          "type": "number"
        },
        "lastRound": {
          "type": "string"
        },
        "foundersNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "keyPeople": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/KeyPeople"
            }
          }
        },
        "source": {
          "type": "string"
        },
        "revenue": {
          "type": "string"
        }
      }
    },
    "Image": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "thumbnail": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
  },
...
  "securityDefinitions": {}
}